requirements.txt
fastapi
uvicorn
httpx
python-dotenv
sentence-transformers
faiss-cpu
numpy